# api/src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Movie Hub
  description: "A Movie Hub API."
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: "Local server"
security:
  - bearerAuth: []
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    page:
      name: page
      in: query
      description: "The page number for pagination."
      schema:
        type: integer
        format: int32
        minimum: 0
        maximum: 10000
        default: 0
        example: 0
    limit:
      name: limit
      in: query
      description: "The maximum number of items to return per page."
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 50
        default: 10
        example: 10
...
    directorId:
      name: directorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89AB][0-9a-f]{3}-[0-9a-f]{12}$"
        maxLength: 255
      description: "The director ID."
      example: f22608f8-43bb-4a69-b77a-cb2e3eaf0dfa
    actorId:
      name: actorId
      in: path
      required: true
      schema:
        type: string
        format: uuid
        pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89AB][0-9a-f]{3}-[0-9a-f]{12}$"
        maxLength: 255
      description: "The actor ID."
      example: c61a0a7c-c45e-41a0-9e6c-3f86beea1802
...
    name:
      name: name
      in: query
      description: "The name of the entity."
      schema:
        type: string
        pattern: "^[A-Za-zÀ-ÿ0-9.,'\"&()\\-: ]*$"
        maxLength: 255
        default: ""
      example: "James Wan"
  responses:
...
    404DirectorNotFound:
      description: Director not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    404ActorNotFound:
      description: Actor not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    401Unauthorized:
      description: Either no access token was provided or the provided access token is invalid or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    403Forbidden:
      description: The user does not have permission to access this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Default:
      description: Unexpected error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    CountryName:
      type: string
      description: |
        The name of the country where the movie was produced or set.
        This field identifies the geographical location associated with the movie and is limited to 15 characters.
      example: Italy
      pattern: "^[A-Za-z\\- ]+$"
      minLength: 1
      maxLength: 15
      nullable: false
...
    MoviePage:
      type: object
      description: "Represents a page of movies, including pagination and sorting information."
      additionalProperties: false
      required:
        - content
        - pageable
        - last
        - totalElements
        - totalPages
        - first
        - size
        - number
        - sort
        - numberOfElements
        - empty
      properties:
        content:
          type: array
          description: "List of movies."
          maxItems: 50
          items:
            $ref: '#/components/schemas/MoviePreviewResponse'
        pageable:
          $ref: '#/components/schemas/PageableDTO'
        last:
          type: boolean
          description: |
            A boolean value indicating whether this is the last page of data.
            If `true`, it is the last page; otherwise, it's not the last page.
          example: false
        totalElements:
          type: integer
          format: int64
          description: "The total number of items."
          example: 4
        totalPages:
          type: integer
          description: "The total number of pages needed to display all the data based on the page size."
          example: 2
        first:
          type: boolean
          description: |
            A boolean value indicating whether this is the first page of data.
            If `true`, it is the first page; otherwise, it's not the first page.
          example: true
        size:
          type: integer
          description: "The maximum number of items contained on each page."
          example: 2
        number:
          type: integer
          description: "The page number for the current page."
          example: 0
        sort:
          $ref: '#/components/schemas/SortDTO'
        numberOfElements:
          type: integer
          description: "The number of items on the current page."
          example: 2
        empty:
          type: boolean
          description: |
            A boolean value indicating whether the current page is empty.
            If `true`, the page has no items; otherwise, it has content.
          example: false
    DirectorPage:
      type: object
      description: "Represents a page of directors, including pagination and sorting information."
      additionalProperties: false
      required:
        - content
        - pageable
        - last
        - totalElements
        - totalPages
        - first
        - size
        - number
        - sort
        - numberOfElements
        - empty
      properties:
        content:
          type: array
          description: "List of directors."
          maxItems: 50
          items:
            $ref: '#/components/schemas/DirectorDetailsResponse'
        pageable:
          $ref: '#/components/schemas/PageableDTO'
        last:
          type: boolean
          description: |
            A boolean value indicating whether this is the last page of data.
            If `true`, it is the last page; otherwise, it's not the last page.
          example: false
        totalElements:
          type: integer
          format: int64
          description: "The total number of items."
          example: 4
        totalPages:
          type: integer
          description: "The total number of pages needed to display all the data based on the page size."
          example: 2
        first:
          type: boolean
          description: |
            A boolean value indicating whether this is the first page of data.
            If `true`, it is the first page; otherwise, it's not the first page.
          example: true
        size:
          type: integer
          description: "The maximum number of items contained on each page."
          example: 2
        number:
          type: integer
          description: "The page number for the current page."
          example: 0
        sort:
          $ref: '#/components/schemas/SortDTO'
        numberOfElements:
          type: integer
          description: "The number of items on the current page."
          example: 2
        empty:
          type: boolean
          description: |
            A boolean value indicating whether the current page is empty.
            If `true`, the page has no items; otherwise, it has content.
          example: false
    ActorPage:
      type: object
      description: "Represents a page of actors, including pagination and sorting information."
      additionalProperties: false
      required:
        - content
        - pageable
        - last
        - totalElements
        - totalPages
        - first
        - size
        - number
        - sort
        - numberOfElements
        - empty
      properties:
        content:
          type: array
          description: "List of actors."
          maxItems: 50
          items:
            $ref: '#/components/schemas/ActorDetailsResponse'
        pageable:
          $ref: '#/components/schemas/PageableDTO'
        last:
          type: boolean
          description: |
            A boolean value indicating whether this is the last page of data.
            If `true`, it is the last page; otherwise, it's not the last page.
          example: false
        totalElements:
          type: integer
          format: int64
          description: "The total number of items."
          example: 4
        totalPages:
          type: integer
          description: "The total number of pages needed to display all the data based on the page size."
          example: 2
        first:
          type: boolean
          description: |
            A boolean value indicating whether this is the first page of data.
            If `true`, it is the first page; otherwise, it's not the first page.
          example: true
        size:
          type: integer
          description: "The maximum number of items contained on each page."
          example: 2
        number:
          type: integer
          description: "The page number for the current page."
          example: 0
        sort:
          $ref: '#/components/schemas/SortDTO'
        numberOfElements:
          type: integer
          description: "The number of items on the current page."
          example: 2
        empty:
          type: boolean
          description: |
            A boolean value indicating whether the current page is empty.
            If `true`, the page has no items; otherwise, it has content.
          example: false
...
    AddDirectorRequest:
      type: object
      additionalProperties: false
      description: |
        Schema for adding a new director to the database.
        It requires the director's name and gender to be provided.
      required:
        - name
        - gender
      properties:
        name:
          $ref: '#/components/schemas/PersonName'
        bio:
          $ref: '#/components/schemas/PersonBio'
        gender:
          $ref: '#/components/schemas/GenderName'
    UpdateDirectorRequest:
      type: object
      additionalProperties: false
      description: |
        Schema for updating an existing director's details.
        It allows modifications to the director's information based on the existing schema.
      properties:
        name:
          $ref: '#/components/schemas/PersonName'
        bio:
          $ref: '#/components/schemas/PersonBio'
        gender:
          $ref: '#/components/schemas/GenderName'
    AddActorRequest:
      type: object
      additionalProperties: false
      description: |
        Schema for adding a new actor to the database.
        It requires the actor's name and gender to be provided.
      required:
        - name
        - gender
      properties:
        name:
          $ref: '#/components/schemas/PersonName'
        bio:
          $ref: '#/components/schemas/PersonBio'
        gender:
          $ref: '#/components/schemas/GenderName'
    UpdateActorRequest:
      type: object
      additionalProperties: false
      description: |
        Schema for updating an existing actor's details.
        It allows changes to the actor's information based on the existing schema.
      properties:
        name:
          $ref: '#/components/schemas/PersonName'
        bio:
          $ref: '#/components/schemas/PersonBio'
        gender:
          $ref: '#/components/schemas/GenderName'
...
    CountryDetailsResponse:
      type: object
      description: "Provides details about a country."
      additionalProperties: false
      required:
        - id
        - name
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/CountryName'
    DirectorDetailsResponse:
      type: object
      description: "Contains detailed information about a director."
      additionalProperties: false
      required:
        - id
        - name
        - gender
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/PersonName'
        bio:
          $ref: '#/components/schemas/PersonBio'
        gender:
          $ref: '#/components/schemas/GenderName'
    ActorDetailsResponse:
      type: object
      description: "Contains detailed information about an actor."
      additionalProperties: false
      required:
        - id
        - name
        - gender
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        name:
          $ref: '#/components/schemas/PersonName'
        bio:
          $ref: '#/components/schemas/PersonBio'
        gender:
          $ref: '#/components/schemas/GenderName'
...
    ProblemDetail:
      type: object
      title: Problem Detail
      description: "A generic error."
      additionalProperties: false
      required:
        - type
        - title
        - status
        - detail
        - instance
      properties:
        type:
          type: string
          description: "A URI reference that identifies the problem type."
          example: "http://localhost:8080/movie-not-found"
          pattern: "^((about:blank)|(https?://[^\\s]+$))"
          maxLength: 1000
        title:
          type: string
          description: "A short, human-readable summary of the problem type."
          example: "Document Not Found"
          pattern: "^[A-Za-z0-9\\s\\-.,!?:;'\"]+$"
          maxLength: 255
        status:
          type: number
          description: "The HTTP status code."
          example: 404
        detail:
          type: string
          description: "A human-readable explanation specific to this occurrence of the problem."
          example: "Movie with ID: 11ae56ed-240e-4d4d-8d3c-f135d9394ab3 not found"
          pattern: "^[A-Za-z0-9\\s\\-.,!?:;'\"]+$"
          maxLength: 1000
        instance:
          type: string
          description: "A URI reference that identifies the specific occurrence of the problem."
          example: "/movies/11ae56ed-240e-4d4d-8d3c-f135d9394ab3"
          pattern: "^(/[^\\s]*)+$"
          maxLength: 1000
        contextInfo:
          type: object
          description: "A free-form based object that contains additional context about the error."
paths:
...
  /directors:
    post:
      description: |
        Create a new director record in the database with the specified details.
        This operation adds a director to the system, enabling it to be used in association with movies.
      summary: Add director
      operationId: addDirector
      security:
        - bearerAuth: []
      tags:
        - directors
      requestBody:
        description: A director to be added to the database.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDirectorRequest'
      responses:
        '201':
          description: Successfully added a director to the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '409':
          description: The director already exists in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        'default':
          $ref: '#/components/responses/Default'
    get:
      description: Retrieve a list of directors available in the system.
      summary: List directors
      operationId: getDirectors
      tags:
        - directors
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/name'
      responses:
        '200':
          description: Successfully retrieved the list of directors available in the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorPage'
        'default':
          $ref: '#/components/responses/Default'
  /directors/{directorId}:
    get:
      description: Retrieve detailed information about a specific director.
      summary: Get director details
      operationId: getDirector
      tags:
        - directors
      parameters:
        - $ref: '#/components/parameters/directorId'
      responses:
        '200':
          description: Successfully retrieved the details of the specified director.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorDetailsResponse'
        '404':
          $ref: '#/components/responses/404DirectorNotFound'
        'default':
          $ref: '#/components/responses/Default'
    put:
      description: |
        Update the details of an existing director.
        All director's details will be replaced with the new details provided.
      summary: Update director details
      operationId: updateDirector
      security:
        - bearerAuth: []
      tags:
        - directors
      parameters:
        - $ref: '#/components/parameters/directorId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDirectorRequest'
      responses:
        '200':
          description: Successfully updated the details of the specified director.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectorDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404DirectorNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /directors/{directorId}/movies:
    get:
      description: Retrieve a list of movies directed by the specified director.
      summary: List movies with director
      operationId: getMoviesWithDirector
      tags:
        - directors
      parameters:
        - $ref: '#/components/parameters/directorId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Successfully retrieved a list of movies directed by the specified director.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePage'
        '404':
          $ref: '#/components/responses/404DirectorNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /actors:
    post:
      description: |
        Create a new actor record in the database with the specified details.
        This operation adds an actor to the system, enabling it to be used in association with movies.
      summary: Add actor
      operationId: addActor
      security:
        - bearerAuth: []
      tags:
        - actors
      requestBody:
        description: An actor to be added to the database.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddActorRequest'
      responses:
        '201':
          description: Successfully added an actor to the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '409':
          description: The actor already exists in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        'default':
          $ref: '#/components/responses/Default'
    get:
      description: Retrieve a list of actors available in the system.
      summary: List actors
      operationId: getActors
      tags:
        - actors
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/name'
      responses:
        '200':
          description: Successfully retrieved the list of actors available in the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorPage'
        'default':
          $ref: '#/components/responses/Default'
  /actors/{actorId}:
    get:
      description: Retrieve detailed information about a specific actor.
      summary: Get actor details
      operationId: getActor
      tags:
        - actors
      parameters:
        - $ref: '#/components/parameters/actorId'
      responses:
        '200':
          description: Successfully retrieved the details of the specified actor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDetailsResponse'
        '404':
          $ref: '#/components/responses/404ActorNotFound'
        'default':
          $ref: '#/components/responses/Default'
    put:
      description: |
        Update the details of an existing actor.
        All actor's details will be replaced with the new details provided.
      summary: Update actor details
      operationId: updateActor
      security:
        - bearerAuth: []
      tags:
        - actors
      parameters:
        - $ref: '#/components/parameters/actorId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateActorRequest'
      responses:
        '200':
          description: Successfully updated the details of the specified actor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404ActorNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /actors/{actorId}/movies:
    get:
      description: Retrieve a list of movies in which the specified actor has appeared.
      summary: List movies with actor
      operationId: getMoviesWithActor
      tags:
        - actors
      parameters:
        - $ref: '#/components/parameters/actorId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Successfully retrieved a list of movies in which the specified actor has appeared.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePage'
        '404':
          $ref: '#/components/responses/404ActorNotFound'
        'default':
          $ref: '#/components/responses/Default'
...
  /countries:
    post:
      description: |
        Create a new country record in the database with the specified details.
        This operation adds a country to the system, enabling it to be used in association with movies.
      summary: Add country
      operationId: addCountry
      security:
        - bearerAuth: []
      tags:
        - countries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: '#/components/schemas/CountryName'
      responses:
        '201':
          description: Successfully created a new country record in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '409':
          description: The country already exists in the database.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        'default':
          $ref: '#/components/responses/Default'
    get:
      description: Retrieve a list of countries available in the system.
      summary: List countries
      operationId: getCountries
      tags:
        - countries
      responses:
        '200':
          description: Successfully retrieved the list of countries available in the system.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryListResponse'
        'default':
          $ref: '#/components/responses/Default'
  /countries/{countryId}:
    get:
      description: Retrieve detailed information about a specific country.
      summary: Get country details
      operationId: getCountry
      tags:
        - countries
      parameters:
        - $ref: '#/components/parameters/countryId'
      responses:
        '200':
          description: Successfully retrieved the details of the specified country.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDetailsResponse'
        '404':
          $ref: '#/components/responses/404CountryNotFound'
        'default':
          $ref: '#/components/responses/Default'
    put:
      description: |
        Update the details of an existing country.
        All country's details will be replaced with the new details provided.
      summary: Update country details
      operationId: updateCountry
      security:
        - bearerAuth: []
      tags:
        - countries
      parameters:
        - $ref: '#/components/parameters/countryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  $ref: '#/components/schemas/CountryName'
      responses:
        '200':
          description: Successfully updated the details of the specified country.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryDetailsResponse'
        '401':
          $ref: '#/components/responses/401Unauthorized'
        '403':
          $ref: '#/components/responses/403Forbidden'
        '404':
          $ref: '#/components/responses/404CountryNotFound'
        'default':
          $ref: '#/components/responses/Default'
  /countries/{countryId}/movies:
    get:
      description: Retrieve a list of movies that were produced in the specified country.
      summary: List movies produced in a country
      operationId: getMoviesWithCountry
      tags:
        - countries
      parameters:
        - $ref: '#/components/parameters/countryId'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
      responses:
        '200':
          description: Successfully retrieved a list of movies that were produced in the specified country.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoviePage'
        '404':
          $ref: '#/components/responses/404CountryNotFound'
        'default':
          $ref: '#/components/responses/Default'